tech debt: fix logic
