+++ .github/codecov.yml
@@ -4,8 +4,8 @@ coverage:
   status:
     project:
       default:
-        target: 90%    # the required coverage value
-        threshold: 3%  # the leniency in hitting the target
+        target: 100%    # the required coverage value
+        threshold: 0.1%  # the leniency in hitting the target
     patch:
       default:
         target: 100%
