+++ src/Cropper.Blazor/Cropper.Blazor.sln
@@ -19,7 +19,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	ProjectSection(SolutionItems) = preProject
 		..\..\.editorconfig = ..\..\.editorconfig
 		..\..\.gitattributes = ..\..\.gitattributes
+		..\..\.github\codecov.yml = ..\..\.github\codecov.yml
 		..\..\.github\CODE_OF_CONDUCT.md = ..\..\.github\CODE_OF_CONDUCT.md
+		..\..\.github\dependabot.yml = ..\..\.github\dependabot.yml
 		..\..\.github\PULL_REQUEST_TEMPLATE.md = ..\..\.github\PULL_REQUEST_TEMPLATE.md
 		..\..\README.md = ..\..\README.md
 	EndProjectSection
