+++ src/Cropper.Blazor/Cropper.Blazor/Components/CropperComponent.razor.Queries.cs
@@ -90,11 +90,15 @@ namespace Cropper.Blazor.Components
         /// Used to get an image from a stream.
         /// <br/>
         /// Converts JavaScript stream to .NET stream uses <see cref="DotNetStreamReference"/> and then creates a URL blob reference.
+        /// <br/>
+        /// Obsolete — use <see cref="Services.IUrlImageInterop.GetImageUsingStreamingAsync(IBrowserFile, long, CancellationToken)"/> 
+        /// for centralized JS interop handling.
         /// </summary>
         /// <param name="imageFile">The <see cref="IBrowserFile"/> to convert to a new image file.</param>
         /// <param name="maxAllowedSize">The maximum number of bytes that can be supplied by the Stream. Defaults to 500 KB.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/> used to propagate notifications that the operation should be canceled.</param>
         /// <returns>A <see cref="ValueTask{String}"/> representing Blob URL asynchronous operation.</returns>
+        [Obsolete("Use IUrlImageInterop.GetImageUsingStreamingAsync instead for centralized JS interop handling.")]
         public async ValueTask<string> GetImageUsingStreamingAsync(
             IBrowserFile imageFile,
             long maxAllowedSize = 512000L,
