+++ src/Cropper.Blazor/Cropper.Blazor/Services/ICropperJsInterop.cs
@@ -6,7 +6,6 @@ using Cropper.Blazor.Base;
 using Cropper.Blazor.Components;
 using Cropper.Blazor.Models;
 using Microsoft.AspNetCore.Components;
-using Microsoft.AspNetCore.Components.Forms;
 using Microsoft.JSInterop;
 
 namespace Cropper.Blazor.Services
@@ -14,7 +13,7 @@ namespace Cropper.Blazor.Services
     /// <summary>
     /// Provides the metadata of a ICropperJsInterop.
     /// </summary>
-    public interface ICropperJsInterop
+    public interface ICropperJsInterop : IUrlImageInterop
     {
         /// <summary>
         /// Loads cropper JS module.
@@ -388,30 +387,6 @@ namespace Cropper.Blazor.Services
             GetCroppedCanvasOptions getCroppedCanvasOptions,
             string type,
             float number,
-            CancellationToken cancellationToken = default);
-
-        /// <summary>
-        /// Used to get an image from a stream.
-        /// <br/>
-        /// Converts JavaScript stream to .NET stream uses <see cref="DotNetStreamReference"/> and then creates a URL blob reference.
-        /// </summary>
-        /// <param name="imageFile">The <see cref="IBrowserFile"/> to convert to a new image file.</param>
-        /// <param name="maxAllowedSize">The maximum number of bytes that can be supplied by the Stream. Defaults to 500 KB.</param>
-        /// <param name="cancellationToken">The <see cref="CancellationToken"/> used to propagate notifications that the operation should be canceled.</param>
-        /// <returns>A <see cref="ValueTask{String}"/> representing URL blob reference asynchronous operation.</returns>
-        ValueTask<string> GetImageUsingStreamingAsync(
-            IBrowserFile imageFile,
-            long maxAllowedSize = 512000L,
-            CancellationToken cancellationToken = default);
-
-        /// <summary>
-        /// Call this method when you've finished using an object URL to let the browser know not to keep the reference to the file any longer.
-        /// </summary>
-        /// <param name="url">A string representing an object URL.</param>
-        /// <param name="cancellationToken">The <see cref="CancellationToken"/> used to propagate notifications that the operation should be canceled.</param>
-        /// <returns>A <see cref="ValueTask"/> representing any asynchronous operation.</returns>
-        ValueTask RevokeObjectUrlAsync(
-            string url,
             CancellationToken cancellationToken = default);
 
         /// <summary>
